Maintenance: Upgrade to Sqlite-3.6.1 and APSW-3.5.9r2
git-svn-id: svn://svn.code.sf.net/p/bitpim/code/trunk/bitpim@4697 4404a10c-ff0d-0410-9031-a1dde6cb222d

+++ src/database.py
@@ -653,7 +653,8 @@ class Database:
         schema=self.getcolumns(tablename)
         
         res=[]
-        for row in self.sqlmany("select * from %s where __rowid__=?" % (idquote(tablename),), [(int(r),) for r in rows.split(',') if len(r)]):
+        for row in self.sqlmany("select * from %s where __rowid__=?" %
+                                (idquote(tablename),), [(int(long(r)),) for r in rows.split(',') if len(r)]):
             record={}
             for colnum,name,type in schema:
                 if name.startswith("__") or type not in ("valueBLOB", "indirectBLOB", "indirectdictBLOB") or row[colnum] is None:
